+++ lib/index.ts
@@ -1,17 +1,24 @@
 import { RushObserver } from "./observer/rush-observer";
 import { RushStream } from "./stream/rush-stream";
 import { RushSubscriber } from "./stream/rush-subscriber";
+import { streamFromEvent, streamFromTarget } from "./utils/from-events";
+
 import type {
   RushDebugHook,
+  RushMiddleware,
   RushObserverImpl,
   RushObserveStream,
   RushUseOption
 } from "./types";
 
-export { RushObserver, RushStream, RushSubscriber };
+
+export {
+  RushObserver, RushStream, RushSubscriber,
+  streamFromEvent, streamFromTarget
+};
 
   export type {
-    RushDebugHook, RushObserverImpl,
+    RushDebugHook, RushMiddleware, RushObserverImpl,
     RushObserveStream, RushUseOption
   };