+++ test/stream/retry.spec.ts
@@ -1,4 +1,4 @@
-import { RushStream } from "../../lib/stream/rush-stream";
+import { RushStream } from "../../lib";
 
 jest.useFakeTimers();
 
@@ -7,7 +7,7 @@ describe("RushStream retry", () => {
     jest.clearAllTimers();
   });
 
-  test("retries stream middleware error", async () => {
+  test("retry stream middleware error", async () => {
     const stream = new RushStream<number>((observer) => {
       observer.next(1);
     }, { continueOnError: true });